requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
streamlit==1.32.0
pdfplumber==0.10.3
pandas==2.2.1
pyarrow==15.0.2
spacy==3.7.4
scikit-learn==1.4.1
prophet==1.1.5
plotly==5.20.0
apscheduler==3.10.4
watchdog==4.0.0
python-dateutil==2.9.0
anthropic==0.21.0
pydantic==2.6.4
sqlalchemy==2.0.28
pytest==8.1.1
faker==24.3.0
reportlab==4.1.0
PyYAML==6.0.2
joblib==1.4.2
ollama
python-dotenv